[agents] New agent 'dispatch' for message routing
