+++ langstream-agents/langstream-agents-commons/src/main/java/ai/langstream/ai/agents/commons/TransformContext.java
@@ -15,6 +15,9 @@
  */
 package ai.langstream.ai.agents.commons;
 
+import ai.langstream.api.runner.code.Header;
+import ai.langstream.api.runner.code.Record;
+import ai.langstream.api.runner.code.SimpleRecord;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.core.type.TypeReference;
 import com.fasterxml.jackson.databind.JsonNode;
@@ -22,13 +25,23 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.node.ObjectNode;
 import java.io.ByteArrayOutputStream;
 import java.io.IOException;
+import java.nio.charset.StandardCharsets;
+import java.sql.Time;
+import java.sql.Timestamp;
+import java.time.Instant;
+import java.time.LocalDate;
+import java.time.LocalDateTime;
+import java.time.LocalTime;
 import java.util.ArrayList;
 import java.util.Collection;
+import java.util.Date;
 import java.util.HashMap;
 import java.util.HashSet;
 import java.util.List;
 import java.util.Map;
+import java.util.Optional;
 import java.util.Set;
+import java.util.stream.Collectors;
 import lombok.Data;
 import lombok.SneakyThrows;
 import lombok.extern.slf4j.Slf4j;
@@ -357,5 +370,173 @@ public class TransformContext {
             return jsonNode.deepCopy();
         }
         throw new UnsupportedOperationException("Cannot copy a value of " + object.getClass());
+    }
+
+    public static TransformContext recordToTransformContext(
+            Record record, boolean attemptJsonConversion) {
+        TransformContext context = new TransformContext();
+        context.setKeyObject(record.key());
+        context.setKeySchemaType(
+                record.key() == null ? null : getSchemaType(record.key().getClass()));
+        // TODO: temporary hack. We should be able to get the schema from the record
+        if (record.key() instanceof GenericRecord) {
+            context.setKeyNativeSchema(((GenericRecord) record.key()).getSchema());
+        }
+        context.setValueObject(record.value());
+        context.setValueSchemaType(
+                record.value() == null ? null : getSchemaType(record.value().getClass()));
+        // TODO: temporary hack. We should be able to get the schema from the record
+        if (record.value() instanceof GenericRecord) {
+            context.setKeyNativeSchema(((GenericRecord) record.value()).getSchema());
+        }
+        context.setInputTopic(record.origin());
+        context.setEventTime(record.timestamp());
+        if (attemptJsonConversion) {
+            context.setKeyObject(attemptJsonConversion(context.getKeyObject()));
+            context.setValueObject(attemptJsonConversion(context.getValueObject()));
+        }
+        // the headers must be Strings, this is a tentative conversion
+        // in the future we need a better way to handle headers
+        context.setProperties(
+                record.headers().stream()
+                        .filter(h -> h.key() != null && h.value() != null)
+                        .collect(
+                                Collectors.toMap(
+                                        Header::key,
+                                        (h -> {
+                                            if (h.value() == null) {
+                                                return null;
+                                            }
+                                            if (h.value() instanceof byte[]) {
+                                                return new String(
+                                                        (byte[]) h.value(), StandardCharsets.UTF_8);
+                                            } else {
+                                                return h.value().toString();
+                                            }
+                                        }))));
+        return context;
+    }
+
+    public static Optional<Record> transformContextToRecord(TransformContext context) {
+        if (context.isDropCurrentRecord()) {
+            return Optional.empty();
+        }
+        List<Header> headers = new ArrayList<>();
+        context.getProperties()
+                .forEach(
+                        (key, value) -> {
+                            SimpleRecord.SimpleHeader header =
+                                    new SimpleRecord.SimpleHeader(key, value);
+                            headers.add(header);
+                        });
+        return Optional.of(new TransformRecord(context, headers));
+    }
+
+    private record TransformRecord(TransformContext context, Collection<Header> headers)
+            implements Record {
+        private TransformRecord(TransformContext context, Collection<Header> headers) {
+            this.context = context;
+            this.headers = new ArrayList<>(headers);
+        }
+
+        @Override
+        public Object key() {
+            return context.getKeyObject();
+        }
+
+        @Override
+        public Object value() {
+            return context.getValueObject();
+        }
+
+        @Override
+        public String origin() {
+            return context.getInputTopic();
+        }
+
+        @Override
+        public Long timestamp() {
+            return context.getEventTime();
+        }
+    }
+
+    private static TransformSchemaType getSchemaType(Class<?> javaType) {
+        if (String.class.isAssignableFrom(javaType)) {
+            return TransformSchemaType.STRING;
+        }
+        if (Byte.class.isAssignableFrom(javaType)) {
+            return TransformSchemaType.INT8;
+        }
+        if (Short.class.isAssignableFrom(javaType)) {
+            return TransformSchemaType.INT16;
+        }
+        if (Integer.class.isAssignableFrom(javaType)) {
+            return TransformSchemaType.INT32;
+        }
+        if (Long.class.isAssignableFrom(javaType)) {
+            return TransformSchemaType.INT64;
+        }
+        if (Double.class.isAssignableFrom(javaType)) {
+            return TransformSchemaType.DOUBLE;
+        }
+        if (Float.class.isAssignableFrom(javaType)) {
+            return TransformSchemaType.FLOAT;
+        }
+        if (Boolean.class.isAssignableFrom(javaType)) {
+            return TransformSchemaType.BOOLEAN;
+        }
+        if (byte[].class.isAssignableFrom(javaType)) {
+            return TransformSchemaType.BYTES;
+        }
+        // Must be before DATE
+        if (Time.class.isAssignableFrom(javaType)) {
+            return TransformSchemaType.TIME;
+        }
+        // Must be before DATE
+        if (Timestamp.class.isAssignableFrom(javaType)) {
+            return TransformSchemaType.TIMESTAMP;
+        }
+        if (Date.class.isAssignableFrom(javaType)) {
+            return TransformSchemaType.DATE;
+        }
+        if (Instant.class.isAssignableFrom(javaType)) {
+            return TransformSchemaType.INSTANT;
+        }
+        if (LocalDate.class.isAssignableFrom(javaType)) {
+            return TransformSchemaType.LOCAL_DATE;
+        }
+        if (LocalTime.class.isAssignableFrom(javaType)) {
+            return TransformSchemaType.LOCAL_TIME;
+        }
+        if (LocalDateTime.class.isAssignableFrom(javaType)) {
+            return TransformSchemaType.LOCAL_DATE_TIME;
+        }
+        if (GenericRecord.class.isAssignableFrom(javaType)) {
+            return TransformSchemaType.AVRO;
+        }
+        if (JsonNode.class.isAssignableFrom(javaType)) {
+            return TransformSchemaType.JSON;
+        }
+        if (Map.class.isAssignableFrom(javaType)) {
+            return TransformSchemaType.MAP;
+        }
+        throw new IllegalArgumentException("Unsupported data type: " + javaType);
+    }
+
+    public static Object attemptJsonConversion(Object value) {
+        try {
+            if (value instanceof String) {
+                return OBJECT_MAPPER.readValue(
+                        (String) value, new TypeReference<Map<String, Object>>() {});
+            } else if (value instanceof byte[]) {
+                return OBJECT_MAPPER.readValue(
+                        (byte[]) value, new TypeReference<Map<String, Object>>() {});
+            }
+        } catch (IOException e) {
+            if (log.isDebugEnabled()) {
+                log.debug("Cannot convert value to json", e);
+            }
+        }
+        return value;
     }
 }
