+++ langstream-agents/langstream-agents-flow-control/src/main/java/ai/langstream/agents/flow/DispatchAgent.java
@@ -1,0 +1,148 @@
+/*
+ * Copyright DataStax, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package ai.langstream.agents.flow;
+
+import ai.langstream.ai.agents.commons.TransformContext;
+import ai.langstream.ai.agents.commons.jstl.predicate.JstlPredicate;
+import ai.langstream.api.runner.code.AbstractAgentCode;
+import ai.langstream.api.runner.code.AgentContext;
+import ai.langstream.api.runner.code.AgentProcessor;
+import ai.langstream.api.runner.code.Record;
+import ai.langstream.api.runner.code.RecordSink;
+import ai.langstream.api.runner.topics.TopicProducer;
+import ai.langstream.api.runtime.ComponentType;
+import ai.langstream.api.util.ConfigurationUtils;
+import java.util.ArrayList;
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
+import lombok.extern.slf4j.Slf4j;
+
+@Slf4j
+public class DispatchAgent extends AbstractAgentCode implements AgentProcessor {
+
+    record Route(String destination, JstlPredicate predicate) {}
+
+    private final List<Route> routes = new ArrayList<>();
+    private final Map<String, TopicProducer> producers = new HashMap<>();
+    private AgentContext agentContext;
+
+    @SuppressWarnings("unchecked")
+    @Override
+    public void init(Map<String, Object> configuration) {
+        List<Map<String, Object>> routes =
+                (List<Map<String, Object>>) configuration.getOrDefault("routes", List.of());
+        routes.forEach(
+                r -> {
+                    String when = ConfigurationUtils.getString("when", "", r);
+                    String destination = ConfigurationUtils.getString("destination", "", r);
+                    if (destination.isEmpty()) {
+                        log.info("Condition: \"{}\", Destination is empty (discard record)", when);
+                    } else {
+                        log.info("Condition: \"{}\", Destination: {}", when, destination);
+                    }
+                    this.routes.add(new Route(destination, new JstlPredicate(when)));
+                });
+    }
+
+    @Override
+    public void setContext(AgentContext context) throws Exception {
+        this.agentContext = context;
+    }
+
+    @Override
+    public void start() throws Exception {
+        routes.forEach(
+                r -> {
+                    String topic = r.destination;
+                    if (topic != null && !topic.isEmpty()) {
+                        TopicProducer producer =
+                                agentContext
+                                        .getTopicConnectionProvider()
+                                        .createProducer(
+                                                agentContext.getGlobalAgentId(), topic, Map.of());
+                        producers.put(topic, producer);
+                    }
+                });
+    }
+
+    @Override
+    public void process(List<Record> records, RecordSink recordSink) {
+        for (Record record : records) {
+            processRecord(record, recordSink);
+        }
+    }
+
+    @Override
+    public ComponentType componentType() {
+        return ComponentType.PROCESSOR;
+    }
+
+    public void processRecord(Record record, RecordSink recordSink) {
+        try {
+            TransformContext context = TransformContext.recordToTransformContext(record, true);
+
+            for (Route r : routes) {
+                boolean test = r.predicate.test(context);
+                if (test) {
+                    if (r.destination.isEmpty()) {
+                        log.info("Discarding record {} - empty destination", record);
+                        recordSink.emit(new SourceRecordAndResult(record, List.of(), null));
+                    } else {
+                        log.info("Dispatching record {} to topic {}", record, r.destination);
+                        TopicProducer topicProducer = producers.get(r.destination);
+                        topicProducer
+                                .write(record)
+                                .whenComplete(
+                                        (__, e) -> {
+                                            if (e != null) {
+                                                log.error(
+                                                        "Error writing record to topic {}",
+                                                        r.destination,
+                                                        e);
+                                                recordSink.emit(
+                                                        new SourceRecordAndResult(record, null, e));
+                                            } else {
+                                                recordSink.emit(
+                                                        new SourceRecordAndResult(
+                                                                record, List.of(record), null));
+                                            }
+                                        });
+                    }
+                    return;
+                }
+            }
+
+            recordSink.emit(new SourceRecordAndResult(record, List.of(), null));
+        } catch (Throwable error) {
+            log.error("Error processing record: {}", record, error);
+            recordSink.emit(new SourceRecordAndResult(record, null, error));
+        }
+    }
+
+    @Override
+    public void close() throws Exception {
+        producers.forEach(
+                (destination, producer) -> {
+                    log.info("Closing producer for destination {}", destination);
+                    try {
+                        producer.close();
+                    } catch (Exception e) {
+                        log.error("Error closing producer for destination {}", destination, e);
+                    }
+                });
+    }
+}
