+++ langstream-agents/langstream-agents-flow-control/src/main/resources/META-INF/ai.langstream.agents.index
@@ -1,0 +1,1 @@
+dispatch
