+++ langstream-agents/langstream-agents-flow-control/src/main/resources/META-INF/services/ai.langstream.api.runner.code.AgentCodeProvider
@@ -1,0 +1,1 @@
+ai.langstream.agents.flow.FlowControlAgentsCodeProvider
