+++ langstream-agents/langstream-ai-agents/src/main/java/ai/langstream/ai/agents/GenAIToolKitAgent.java
@@ -15,17 +15,17 @@
  */
 package ai.langstream.ai.agents;
 
+import static ai.langstream.ai.agents.commons.TransformContext.recordToTransformContext;
+import static ai.langstream.ai.agents.commons.TransformContext.transformContextToRecord;
+
 import ai.langstream.ai.agents.commons.TransformContext;
-import ai.langstream.ai.agents.commons.TransformSchemaType;
 import ai.langstream.ai.agents.datasource.DataSourceProviderRegistry;
 import ai.langstream.ai.agents.services.ServiceProviderRegistry;
 import ai.langstream.api.runner.code.AbstractAgentCode;
 import ai.langstream.api.runner.code.AgentContext;
 import ai.langstream.api.runner.code.AgentProcessor;
-import ai.langstream.api.runner.code.Header;
 import ai.langstream.api.runner.code.Record;
 import ai.langstream.api.runner.code.RecordSink;
-import ai.langstream.api.runner.code.SimpleRecord;
 import ai.langstream.api.runner.topics.TopicProducer;
 import ai.langstream.api.runtime.ComponentType;
 import com.datastax.oss.streaming.ai.StepPredicatePair;
@@ -37,18 +37,7 @@ import com.datastax.oss.streaming.ai.services.ServiceProvider;
 import com.datastax.oss.streaming.ai.streaming.StreamingAnswersConsumer;
 import com.datastax.oss.streaming.ai.streaming.StreamingAnswersConsumerFactory;
 import com.datastax.oss.streaming.ai.util.TransformFunctionUtil;
-import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
-import java.nio.charset.StandardCharsets;
-import java.sql.Time;
-import java.sql.Timestamp;
-import java.time.Instant;
-import java.time.LocalDate;
-import java.time.LocalDateTime;
-import java.time.LocalTime;
-import java.util.ArrayList;
-import java.util.Collection;
-import java.util.Date;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
@@ -56,10 +45,8 @@ import java.util.Optional;
 import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.CompletionException;
 import java.util.function.Predicate;
-import java.util.stream.Collectors;
 import lombok.SneakyThrows;
 import lombok.extern.slf4j.Slf4j;
-import org.apache.avro.generic.GenericRecord;
 
 @Slf4j
 public class GenAIToolKitAgent extends AbstractAgentCode implements AgentProcessor {
@@ -191,159 +178,6 @@ public class GenAIToolKitAgent extends AbstractAgentCode implements AgentProcess
         }
     }
 
-    public static TransformContext recordToTransformContext(
-            Record record, boolean attemptJsonConversion) {
-        TransformContext context = new TransformContext();
-        context.setKeyObject(record.key());
-        context.setKeySchemaType(
-                record.key() == null ? null : getSchemaType(record.key().getClass()));
-        // TODO: temporary hack. We should be able to get the schema from the record
-        if (record.key() instanceof GenericRecord) {
-            context.setKeyNativeSchema(((GenericRecord) record.key()).getSchema());
-        }
-        context.setValueObject(record.value());
-        context.setValueSchemaType(
-                record.value() == null ? null : getSchemaType(record.value().getClass()));
-        // TODO: temporary hack. We should be able to get the schema from the record
-        if (record.value() instanceof GenericRecord) {
-            context.setKeyNativeSchema(((GenericRecord) record.value()).getSchema());
-        }
-        context.setInputTopic(record.origin());
-        context.setEventTime(record.timestamp());
-        if (attemptJsonConversion) {
-            context.setKeyObject(
-                    TransformFunctionUtil.attemptJsonConversion(context.getKeyObject()));
-            context.setValueObject(
-                    TransformFunctionUtil.attemptJsonConversion(context.getValueObject()));
-        }
-        // the headers must be Strings, this is a tentative conversion
-        // in the future we need a better way to handle headers
-        context.setProperties(
-                record.headers().stream()
-                        .filter(h -> h.key() != null && h.value() != null)
-                        .collect(
-                                Collectors.toMap(
-                                        Header::key,
-                                        (h -> {
-                                            if (h.value() == null) {
-                                                return null;
-                                            }
-                                            if (h.value() instanceof byte[]) {
-                                                return new String(
-                                                        (byte[]) h.value(), StandardCharsets.UTF_8);
-                                            } else {
-                                                return h.value().toString();
-                                            }
-                                        }))));
-        return context;
-    }
-
-    public static Optional<Record> transformContextToRecord(TransformContext context) {
-        if (context.isDropCurrentRecord()) {
-            return Optional.empty();
-        }
-        List<Header> headers = new ArrayList<>();
-        context.getProperties()
-                .forEach(
-                        (key, value) -> {
-                            SimpleRecord.SimpleHeader header =
-                                    new SimpleRecord.SimpleHeader(key, value);
-                            headers.add(header);
-                        });
-        return Optional.of(new TransformRecord(context, headers));
-    }
-
-    private record TransformRecord(TransformContext context, Collection<Header> headers)
-            implements Record {
-        private TransformRecord(TransformContext context, Collection<Header> headers) {
-            this.context = context;
-            this.headers = new ArrayList<>(headers);
-        }
-
-        @Override
-        public Object key() {
-            return context.getKeyObject();
-        }
-
-        @Override
-        public Object value() {
-            return context.getValueObject();
-        }
-
-        @Override
-        public String origin() {
-            return context.getInputTopic();
-        }
-
-        @Override
-        public Long timestamp() {
-            return context.getEventTime();
-        }
-    }
-
-    private static TransformSchemaType getSchemaType(Class<?> javaType) {
-        if (String.class.isAssignableFrom(javaType)) {
-            return TransformSchemaType.STRING;
-        }
-        if (Byte.class.isAssignableFrom(javaType)) {
-            return TransformSchemaType.INT8;
-        }
-        if (Short.class.isAssignableFrom(javaType)) {
-            return TransformSchemaType.INT16;
-        }
-        if (Integer.class.isAssignableFrom(javaType)) {
-            return TransformSchemaType.INT32;
-        }
-        if (Long.class.isAssignableFrom(javaType)) {
-            return TransformSchemaType.INT64;
-        }
-        if (Double.class.isAssignableFrom(javaType)) {
-            return TransformSchemaType.DOUBLE;
-        }
-        if (Float.class.isAssignableFrom(javaType)) {
-            return TransformSchemaType.FLOAT;
-        }
-        if (Boolean.class.isAssignableFrom(javaType)) {
-            return TransformSchemaType.BOOLEAN;
-        }
-        if (byte[].class.isAssignableFrom(javaType)) {
-            return TransformSchemaType.BYTES;
-        }
-        // Must be before DATE
-        if (Time.class.isAssignableFrom(javaType)) {
-            return TransformSchemaType.TIME;
-        }
-        // Must be before DATE
-        if (Timestamp.class.isAssignableFrom(javaType)) {
-            return TransformSchemaType.TIMESTAMP;
-        }
-        if (Date.class.isAssignableFrom(javaType)) {
-            return TransformSchemaType.DATE;
-        }
-        if (Instant.class.isAssignableFrom(javaType)) {
-            return TransformSchemaType.INSTANT;
-        }
-        if (LocalDate.class.isAssignableFrom(javaType)) {
-            return TransformSchemaType.LOCAL_DATE;
-        }
-        if (LocalTime.class.isAssignableFrom(javaType)) {
-            return TransformSchemaType.LOCAL_TIME;
-        }
-        if (LocalDateTime.class.isAssignableFrom(javaType)) {
-            return TransformSchemaType.LOCAL_DATE_TIME;
-        }
-        if (GenericRecord.class.isAssignableFrom(javaType)) {
-            return TransformSchemaType.AVRO;
-        }
-        if (JsonNode.class.isAssignableFrom(javaType)) {
-            return TransformSchemaType.JSON;
-        }
-        if (Map.class.isAssignableFrom(javaType)) {
-            return TransformSchemaType.MAP;
-        }
-        throw new IllegalArgumentException("Unsupported data type: " + javaType);
-    }
-
     private static class TopicProducerStreamingAnswersConsumerFactory
             implements StreamingAnswersConsumerFactory {
         private AgentContext agentContext;
@@ -355,8 +189,7 @@ public class GenAIToolKitAgent extends AbstractAgentCode implements AgentProcess
             TopicProducer topicProducer =
                     agentContext
                             .getTopicConnectionProvider()
-                            .createProducer(
-                                    agentContext.getGlobalAgentId(), Map.of("topic", topicName));
+                            .createProducer(agentContext.getGlobalAgentId(), topicName, Map.of());
             topicProducer.start();
             return new TopicStreamingAnswersConsumer(topicProducer);
         }
