+++ langstream-agents/langstream-vector-agents/src/main/java/ai/langstream/agents/vector/pinecone/PineconeWriter.java
@@ -15,7 +15,8 @@
  */
 package ai.langstream.agents.vector.pinecone;
 
-import ai.langstream.ai.agents.GenAIToolKitAgent;
+import static ai.langstream.ai.agents.commons.TransformContext.recordToTransformContext;
+
 import ai.langstream.ai.agents.commons.TransformContext;
 import ai.langstream.ai.agents.commons.jstl.JstlEvaluator;
 import ai.langstream.api.database.VectorDatabaseWriter;
@@ -102,8 +103,7 @@ public class PineconeWriter implements VectorDatabaseWriterProvider {
         public CompletableFuture<?> upsert(Record record, Map<String, Object> context) {
             CompletableFuture<?> handle = new CompletableFuture<>();
             try {
-                TransformContext transformContext =
-                        GenAIToolKitAgent.recordToTransformContext(record, true);
+                TransformContext transformContext = recordToTransformContext(record, true);
                 String id =
                         idFunction != null ? (String) idFunction.evaluate(transformContext) : null;
                 String namespace =
